# internal/docs/swagger.yaml
basePath: /
definitions:
  aclmodels.Acl2Scope:
    enum:
    - ""
    - ror
    - cluster
    - project
    - datacenter
    - virtualmachine
    type: string
    x-enum-comments:
      Acl2ScopeRor: ROR
      Acl2ScopeUnknown: unknown
    x-enum-varnames:
    - Acl2ScopeUnknown
    - Acl2ScopeRor
    - Acl2ScopeCluster
    - Acl2ScopeProject
    - Acl2ScopeDatacenter
    - Acl2ScopeVirtualMachine
  aclmodels.AclV2ListItem:
    properties:
      access:
        allOf:
        - $ref: '#/definitions/aclmodels.AclV2ListItemAccess'
        description: v2 access model for ror api
      created:
        type: string
      group:
        description: The group which the acces is granted
        minLength: 1
        type: string
      id:
        description: Id
        type: string
      issuedBy:
        type: string
      kubernetes:
        allOf:
        - $ref: '#/definitions/aclmodels.AclV2ListItemKubernetes'
        description: Accessv2   []map[AccessType]bool    `json:"accessv2" validate:""`                      //
          v2 access model for ror api
      scope:
        allOf:
        - $ref: '#/definitions/aclmodels.Acl2Scope'
        description: Type of object ['cluster','project']
        minLength: 1
      subject:
        description: The subject eg. clusterid, projectid (can be 'All')
        minLength: 1
        type: string
      version:
        default: 2
        description: Acl Version, must be 2
        type: integer
    required:
    - access
    - group
    - scope
    - subject
    type: object
  aclmodels.AclV2ListItemAccess:
    properties:
      create:
        description: Write metadata of subject
        type: boolean
      delete:
        description: Delete metadata of subject
        type: boolean
      owner:
        description: Delete metadata of subject
        type: boolean
      read:
        description: Read metadata of subject
        type: boolean
      update:
        description: Update metadata of subject
        type: boolean
    type: object
  aclmodels.AclV2ListItemKubernetes:
    properties:
      logon:
        description: Logon to subject if 'cluster'
        type: boolean
    type: object
...
  apicontracts.Filter:
    properties:
      filters:
        items:
          $ref: '#/definitions/apicontracts.FilterMetadata'
        type: array
      globalFilter: {}
      limit:
        type: integer
      skip:
        type: integer
      sort:
        items:
          $ref: '#/definitions/apicontracts.SortMetadata'
        type: array
    type: object
  apicontracts.FilterMetadata:
    properties:
      field:
        type: string
      matchMode:
        $ref: '#/definitions/apicontracts.MatchModeType'
      value: {}
    type: object
...
  apicontracts.MatchModeType:
    enum:
    - unknown
    - contains
    - equals
    - in
    type: string
    x-enum-varnames:
    - MatchModeUnknown
    - MatchModeContains
    - MatchModeEquals
    - MatchModeIn
...
  apicontracts.PaginatedResult-aclmodels_AclV2ListItem:
    properties:
      data:
        items:
          $ref: '#/definitions/aclmodels.AclV2ListItem'
        type: array
      dataCount:
        type: integer
      offset:
        type: integer
      totalCount:
        type: integer
    type: object
...
  apicontracts.SortMetadata:
    properties:
      sortField:
        type: string
      sortOrder:
        type: integer
    type: object
...
  rorerror.RorError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    name: ROR
    url: https://github.com/NorskHelsenett/ror
  description: 'ROR-API, need any help? Go to channel #drift-sdi-devops in norskhelsenett.slack.com
    slack workspace'
  title: Swagger ROR-API
  version: "0.1"
paths:
...
  /v1/acl:
    post:
      consumes:
      - application/json
      description: Create acl
      parameters:
      - description: Acl
        in: body
        name: acl
        required: true
        schema:
          $ref: '#/definitions/aclmodels.AclV2ListItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aclmodels.AclV2ListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Create acl
      tags:
      - acl
  /v1/acl/{aclId}:
    delete:
      consumes:
      - application/json
      description: Delete a acl by id
      parameters:
      - description: aclId
        in: path
        name: aclId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: bool
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Delete acl
      tags:
      - acl
    get:
      consumes:
      - application/json
      description: Get acl by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.PaginatedResult-aclmodels_AclV2ListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get acl by id
      tags:
      - acl
    put:
      consumes:
      - application/json
      description: Update acl
      parameters:
      - description: aclId
        in: path
        name: aclId
        required: true
        type: string
      - description: Acl
        in: body
        name: acl
        required: true
        schema:
          $ref: '#/definitions/aclmodels.AclV2ListItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aclmodels.AclV2ListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Update acl
      tags:
      - acl
...
  /v1/acl/filter:
    post:
      consumes:
      - application/json
      description: Get acl by filter
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/apicontracts.Filter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apicontracts.PaginatedResult-aclmodels_AclV2ListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get acl by filter
      tags:
      - acl
  /v1/acl/migrate:
    get:
      consumes:
      - application/json
      description: Migrate acl
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - AccessToken: []
        ApiKey: []
      summary: Migrate acl
      tags:
      - acl
  /v1/acl/scopes:
    get:
      consumes:
      - application/json
      description: Get acl scopes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/aclmodels.Acl2Scope'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rorerror.RorError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rorerror.RorError'
      security:
      - AccessToken: []
        ApiKey: []
      summary: Get acl scopes
      tags:
      - acl
...
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
  ApiKey:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"